Update SMDiskStore.pm
Wolfgang Baudler <wbaudler@gb.nrao.edu> reports that email bodies which do not end with an empty line cause SpamAssassin to miss rules that should hit. In his installation, ensuring that the results of ReadBody() always end with a newline fixes the issue.

I don't have a sendmail test environment. I can't duplicate the issue with Postfix, and I can't test this fix, but I *think* it should be OK. It should be tested in a sendmail environment before merging.